exception handled both custom and global,used loggers,created custom logic for account number generation and used validation

+++ AccountService/src/main/java/com/maveric/AccountService/payload/CurrentAccountDto.java
@@ -1,8 +1,11 @@
 package com.maveric.AccountService.payload;
 
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.validator.constraints.Length;
 
 @Data
 @NoArgsConstructor
@@ -11,7 +14,11 @@ public class CurrentAccountDto {
 
     private long id;
     private float balance;
-    private int accountnumber;
+
+    private long accountnumber;
+
+
+
 }
